Basic implementaiton
The basic implementation of this feature has a bug. The virtual mouse and the mouse can not be active at the same time at present.

The implementation is there, but it needs some kind of workaround.

- Eventsystem has a new source
- A cursor object is appended to the building hud object
-The cursor object has a virtual mouse script

I saved Unitys Readme in the folder default scripts, where the virtual mouse script is.